Implement basic VN message box with coroutines
Walk through the plot using Lua code contained in a coroutine.
Will allow branching and simmilar flow control using existing
Lua sintax, and this should in theory be simpler than having
a custom scripting language on top of the existing one.

Known issue:
Last message requires a second button press before it goes away.
This is due to the coroutine yielding, and should be simple to fix.