Treat plot as an itterator function
Plot is a itterator function that as long as there is something to show
returns the next thing to show, and nil once it's depleated.

This should make it usable in a for each loop (if that's ever needed).

Also, minimal logging infrastructure in place.

Bonus: Eccentric function naming convention came to me in a (day-)dream.