Support multiple parallell elements in VN UI
Allow plot to queue up multiple UI elements that are then displayed
at the same time. This allows for somewhat more complex pressentations.

Also, go nuts with the new eccentric function and field naming style.
So far it's proving to be quite nice, and probably ticks of almost
anyone with nothing better to do than have opinions about this stuff.